requirements.txt
beautifulsoup4==4.4.1
cssselect==0.9.1
lxml==3.4.4
requests==2.7.0
wheel==0.24.0
PySocks